Add files via upload
This is an alien invasion game where a ship fires and brings down the ship. There is a constant increasing level if the user plays well and a high score is maintained. The game will run in full screen and the speed of every aspect can be changed.
The game can be run by running the alien_invasion.py file.